add smarten punctuation to PlainTextEdit
