+++ src/calibre/gui2/tweak_book/widgets.py
@@ -1486,9 +1486,9 @@ class PlainTextEdit(QPlainTextEdit):  # {{{
     """A class that overrides some methods from QPlainTextEdit to fix handling
     of the nbsp unicode character and AltGr input method on windows."""
 
-    def __init__(self, parent=None):
+    def __init__(self, parent=None, *, use_smarten_punctuation: bool = False):
         QPlainTextEdit.__init__(self, parent)
-        self.syntax = None
+        self.use_smarten_punctuation = use_smarten_punctuation
 
     def toPlainText(self):
         return to_plain_text(self)
@@ -1563,6 +1563,22 @@ class PlainTextEdit(QPlainTextEdit):  # {{{
                 break
         if changed:
             self.setTextCursor(c)
+
+    def contextMenuEvent(self, e):
+        m = self.createStandardContextMenu()
+        assert m is not None
+        if self.use_smarten_punctuation:
+            m.addSeparator()
+            m.addAction(_('Smarten punctuation'), self.smarten_punctuation)
+        m.exec(e.globalPos())
+
+    def smarten_punctuation(self):
+        from calibre.ebooks.conversion.preprocess import smarten_punctuation
+
+        text = self.toPlainText().strip()
+        newtext = smarten_punctuation(text)
+        if text != newtext:
+            self.setPlainText(newtext)
 
 
 # }}}
